First draft of print_alloc_memory

+++ libft/libft.h
@@ -48,7 +48,7 @@ char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void		ft_striteri(char *s, void (*f)(unsigned int, char*));
 void		ft_putchar_fd(char c, int fd);
 void		ft_putendl_fd(char *s, int fd);
-void		ft_putnbr_fd(int n, int fd);
+void		ft_putnbr_fd(long long n, int fd);
 void		ft_putstr_fd(char *s, int fd);
 typedef struct s_list
 {
